host_driver/requirements.txt
pymongo==4.8.0
pyusb==1.2.1
pyyaml==6.0.1
numpy==1.26.4
progress==1.6
certifi==2024.7.4
